[INJIMOB-1771]: Moving methods from PreProcessor class to Utils and MultilineHanlderTest for modularization
